Overhauled GUI Code and Adding a Record Works
Removed a lot of code from a bunch of files that didn't involve using the GUI at all. Stopped trying to code the entire GUI manually and began using a JFrame Form. Records can now be added to the Database effortlessly and even denies adding a new record if the required fields (first name, last name, and age) are empty.